# apps/api/openapi.json
{
  "components": {
    "schemas": {
      "ActionKind": {
        "description": "What the planner can decide to do.",
        "enum": ["reorder", "dispatch", "notify", "escalate"],
        "title": "ActionKind",
        "type": "string"
      },
      "ActionOut": {
        "description": "A planner decision and its review state.",
        "properties": {
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "estimated_cost_kobo": {
            "title": "Estimated Cost Kobo",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/ActionKind"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "rationale": {
            "title": "Rationale",
            "type": "string"
          },
          "requires_review": {
            "title": "Requires Review",
            "type": "boolean"
          },
          "review_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Note"
          },
          "review_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Reason"
          },
          "reviewed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed At"
          },
          "reviewed_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed By"
          },
          "run_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "status": {
            "$ref": "#/components/schemas/ActionStatus"
          },
          "store_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Store Id"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "run_id",
          "store_id",
          "kind",
          "status",
          "requires_review",
          "review_reason",
          "payload",
          "estimated_cost_kobo",
          "confidence",
          "rationale",
          "reviewed_by",
          "reviewed_at",
          "review_note",
          "created_at",
          "updated_at"
        ],
        "title": "ActionOut",
        "type": "object"
      },
      "ActionStatus": {
        "description": "Lifecycle of a planner decision.",
        "enum": ["proposed", "pending_review", "approved", "rejected", "executed"],
        "title": "ActionStatus",
        "type": "string"
      },
...
      "ErrorResponse": {
        "description": "Body of every 4xx/5xx the API raises deliberately.",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": ["detail"],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RunOut": {
        "description": "An agent run with its tool calls and actions: the timeline the dashboard shows.",
        "properties": {
          "actions": {
            "items": {
              "$ref": "#/components/schemas/ActionOut"
            },
            "title": "Actions",
            "type": "array"
          },
          "attempts": {
            "title": "Attempts",
            "type": "integer"
          },
          "cache_read_tokens": {
            "title": "Cache Read Tokens",
            "type": "integer"
          },
          "cache_write_tokens": {
            "title": "Cache Write Tokens",
            "type": "integer"
          },
          "cost_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Usd"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "extraction_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extraction Id"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "input_tokens": {
            "title": "Input Tokens",
            "type": "integer"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "latency_ms": {
            "title": "Latency Ms",
            "type": "integer"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "output_tokens": {
            "title": "Output Tokens",
            "type": "integer"
          },
          "photo_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Id"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "summary": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "tool_calls": {
            "items": {
              "$ref": "#/components/schemas/ToolCallOut"
            },
            "title": "Tool Calls",
            "type": "array"
          },
          "trigger_role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trigger Role"
          }
        },
        "required": [
          "id",
          "kind",
          "status",
          "trigger_role",
          "photo_id",
          "extraction_id",
          "provider",
          "model",
          "input_tokens",
          "output_tokens",
          "cache_read_tokens",
          "cache_write_tokens",
          "cost_usd",
          "latency_ms",
          "attempts",
          "error",
          "summary",
          "started_at",
          "finished_at",
          "tool_calls",
          "actions"
        ],
        "title": "RunOut",
        "type": "object"
      },
      "RunQueued": {
        "description": "Acknowledgement that a planner run was queued.",
        "properties": {
          "extraction_id": {
            "format": "uuid",
            "title": "Extraction Id",
            "type": "string"
          },
          "queued": {
            "default": true,
            "title": "Queued",
            "type": "boolean"
          }
        },
        "required": ["extraction_id"],
        "title": "RunQueued",
        "type": "object"
      },
      "RunStatus": {
        "description": "Lifecycle of an agent run.",
        "enum": ["queued", "running", "succeeded", "failed"],
        "title": "RunStatus",
        "type": "string"
      },
...
      "ToolCallOut": {
        "description": "One logged tool call.",
        "properties": {
          "arguments": {
            "additionalProperties": true,
            "title": "Arguments",
            "type": "object"
          },
          "caller_role": {
            "title": "Caller Role",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "duration_ms": {
            "title": "Duration Ms",
            "type": "integer"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "seq": {
            "title": "Seq",
            "type": "integer"
          },
          "tool_name": {
            "title": "Tool Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "seq",
          "tool_name",
          "caller_role",
          "arguments",
          "result",
          "error",
          "duration_ms",
          "created_at"
        ],
        "title": "ToolCallOut",
        "type": "object"
      },
      "ToolRunOut": {
        "description": "Result of a direct tool call.",
        "properties": {
          "duration_ms": {
            "title": "Duration Ms",
            "type": "integer"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          }
        },
        "required": ["name", "result", "error", "duration_ms"],
        "title": "ToolRunOut",
        "type": "object"
      },
      "ToolSpecOut": {
        "description": "A tool the caller may invoke, with its JSON Schema.",
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "input_schema": {
            "additionalProperties": true,
            "title": "Input Schema",
            "type": "object"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": ["name", "description", "input_schema"],
        "title": "ToolSpecOut",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": ["loc", "msg", "type"],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Shelf photos and cold-chain telemetry in, reviewed decisions out. The contract is written first in apps/api/openapi.yaml; this document is generated from the implementation and checked against it in CI.",
    "title": "ShelfSense API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/extractions/{extraction_id}/plan": {
      "post": {
        "description": "Queue a planner run over an existing extraction (it also runs automatically).",
        "operationId": "plan_extraction_v1_extractions__extraction_id__plan_post",
        "parameters": [
          {
            "in": "path",
            "name": "extraction_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Extraction Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunQueued"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not authenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not allowed"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Plan Extraction",
        "tags": ["runs"]
      }
    },
...
    "/v1/runs": {
      "get": {
        "description": "Most recent runs first, optionally by kind or status.",
        "operationId": "list_runs_v1_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 32,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/RunStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RunOut"
                  },
                  "title": "Response List Runs V1 Runs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not authenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not allowed"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Runs",
        "tags": ["runs"]
      }
    },
    "/v1/runs/{run_id}": {
      "get": {
        "description": "One run with every tool call and action.",
        "operationId": "read_run_v1_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not authenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not allowed"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Run",
        "tags": ["runs"]
      }
    },
...
    "/v1/tools": {
      "get": {
        "description": "Tools the caller's role may invoke, with their JSON Schemas.",
        "operationId": "list_tools_v1_tools_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ToolSpecOut"
                  },
                  "title": "Response List Tools V1 Tools Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not authenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not allowed"
          }
        },
        "summary": "List Tools",
        "tags": ["tools"]
      }
    },
    "/v1/tools/{name}": {
      "post": {
        "description": "Execute one tool under the caller's tenant and role. Logged like any planner call.",
        "operationId": "run_tool_v1_tools__name__post",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Arguments",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolRunOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not authenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not allowed"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Tool",
        "tags": ["tools"]
      }
    }
  },
...
}